+++ packages/site/src/providers/AtlantisPreviewEditorProvider.tsx
@@ -15,6 +15,7 @@ import {
   defaultHighlightStyle,
   syntaxHighlighting,
 } from "@codemirror/language";
+import { Theme, useAtlantisTheme } from "@jobber/components";
 
 /***
  *
@@ -49,14 +50,18 @@ export const useAtlantisPreview = () => {
 export const AtlantisPreviewEditorProvider = ({
   children,
 }: PropsWithChildren) => {
+  const { theme } = useAtlantisTheme();
   const iframe = useRef<HTMLIFrameElement>(null);
   const [code, setCode] = useState<string>("");
   const [error, setError] = useState<string>("");
 
-  const writeSkeleton = (doc: Document | null | undefined) => {
+  const writeSkeleton = (
+    doc: Document | null | undefined,
+    iframeTheme: Theme,
+  ) => {
     if (doc) {
       doc.open();
-      doc.write(skeletonHTML);
+      doc.write(skeletonHTML(iframeTheme));
       doc.close();
     }
   };
@@ -80,7 +85,7 @@ export const AtlantisPreviewEditorProvider = ({
         const html = iframe.current?.contentDocument?.documentElement.outerHTML;
 
         if (html === "<html><head></head><body></body></html>") {
-          writeSkeleton(iframe.current?.contentDocument);
+          writeSkeleton(iframe.current?.contentDocument, theme);
         }
 
         if (iframe.current) {
@@ -91,6 +96,7 @@ export const AtlantisPreviewEditorProvider = ({
             import {
               AnimatedPresence,
               AnimatedSwitcher,
+              AtlantisThemeContextProvider,
               Autocomplete,
               Avatar,
               Banner,
@@ -186,6 +192,9 @@ export const AtlantisPreviewEditorProvider = ({
 
                 ${transpiledCode}
 
+            function RootWrapper() {
+              return React.createElement(AtlantisThemeContextProvider, null, React.createElement(App));
+            }
 
           if (rootElement) {
               ReactDOM.unmountComponentAtNode(rootElement);
@@ -194,7 +203,7 @@ export const AtlantisPreviewEditorProvider = ({
               rootElement = document.getElementById('root')
               root = ReactDOM.createRoot(rootElement);
              }
-              root.render(React.createElement(App, null));
+              root.render(React.createElement(RootWrapper, null));
           `;
             iframeWindow.postMessage(
               { type: "updateCode", code: codeWrapper },
@@ -208,8 +217,15 @@ export const AtlantisPreviewEditorProvider = ({
         setError((e as { message: string }).message as string);
       }
     },
-    [iframe],
+    [iframe, theme],
   );
+
+  useEffect(() => {
+    if (iframe.current) {
+      const iframeWindow = iframe.current.contentWindow;
+      iframeWindow?.postMessage({ type: "updateTheme", theme }, "*");
+    }
+  }, [theme]);
 
   return (
     <AtlantisPreviewEditorContext.Provider
@@ -300,10 +316,9 @@ export const AtlantisPreviewEditor = () => {
   );
 };
 
-const skeletonHTML = `
-
+const skeletonHTML = (theme: Theme) => `
 <!DOCTYPE html>
-<html>
+<html data-theme="${theme}">
 <head>
 <style>
 html,body,#root {
@@ -318,6 +333,7 @@ html,body,#root {
 </style>
 <link rel="stylesheet" href="/styles.css">
 <link rel="stylesheet" href="/foundation.css">
+<link rel="stylesheet" href="/dark.mode.css">
 </head>
       <body>
  <script type="importmap">
@@ -343,7 +359,8 @@ html,body,#root {
         return true;
       };
       window.addEventListener('message', (event) => {
-        const { type, code } = event.data;
+        const { type, code, theme } = event.data;
+
         if (type === 'updateCode') {
           const script = document.createElement('script');
           script.type = 'module';
@@ -352,6 +369,8 @@ html,body,#root {
           if (root) {
             root.appendChild(script); // Inject new script
           }
+        } else if (type === 'updateTheme') {
+          document.documentElement.dataset.theme = theme;
         }
       });
       </script>