+++ packages/site/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.13.2](https://github.com/GetJobber/atlantis/compare/@jobber/atlantis-site@0.13.1...@jobber/atlantis-site@0.13.2) (2024-12-04)
+
+**Note:** Version bump only for package @jobber/atlantis-site
+
+
+
+
+
+## [0.13.1](https://github.com/GetJobber/atlantis/compare/@jobber/atlantis-site@0.13.0...@jobber/atlantis-site@0.13.1) (2024-12-04)
+
+**Note:** Version bump only for package @jobber/atlantis-site
+
+
+
+
+
 # [0.13.0](https://github.com/GetJobber/atlantis/compare/@jobber/atlantis-site@0.12.7...@jobber/atlantis-site@0.13.0) (2024-12-04)